+++ KeyboardKitDemoKeyboard/DemoKeyboardActionHandler.swift
@@ -40,14 +40,14 @@ class DemoKeyboardActionHandler: StandardKeyboardActionHandler {
     
     // MARK: - Actions
     
-    override func doubleTapAction(for action: KeyboardAction, sender: Any?) -> StandardKeyboardActionHandler.GestureAction? {
+    override func doubleTapAction(for action: KeyboardAction, sender: Any?) -> GestureAction? {
         switch action {
         case .shift: return switchToCapsLockedKeyboard
         default: return super.doubleTapAction(for: action, sender: sender)
         }
     }
     
-    override func longPressAction(for action: KeyboardAction, sender: Any?) -> StandardKeyboardActionHandler.GestureAction? {
+    override func longPressAction(for action: KeyboardAction, sender: Any?) -> GestureAction? {
         switch action {
         case .image(_, _, let imageName): return { [weak self] in self?.saveImage(UIImage(named: imageName)!) }
         case .shift: return switchToCapsLockedKeyboard
@@ -151,6 +151,7 @@ private extension DemoKeyboardActionHandler {
     }
     
     func switchToUppercaseKeyboard() {
+        if keyboardShiftState == .capsLocked { return }
         switchToAlphabeticKeyboard( .uppercased)
     }
 }